pyproject.toml
[project]
name = "wsi-deidentifier"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10.0"
dependencies = [
  "fastapi[all]>=0.115.12",
  "google>=3.0.0",
  "google-cloud-vision>=3.10.1",
  "google-genai>=1.12.1",
  "matplotlib>=3.10.1",
  "opencv-python>=4.11.0.86",
  "openslide-bin>=4.0.0.8",
  "openslide-python>=1.4.1",
  "pillow>=11.2.1",
  "pylibdmtx>=0.1.10",
  "pytesseract>=0.3.13",
  "python-dotenv>=1.1.0",
  "pyzbar>=0.1.9",
  "setuptools>=80.1.0",
  "tifffile>=2025.3.30",
  "tinynumpy>=1.2.1",
]

[tool.uv]
